Big reorganise part 2
