+++ src/components/Routes/index.js
@@ -1,0 +1,2 @@
+export { default as Start } from './Start';
+export { default as Protocol } from './Protocol';
