+++ src/components/Rule/index.js
@@ -1,1 +1,4 @@
 export { default } from './Rule';
+export { default as DropDown } from './DropDown';
+export { default as AddButton } from './AddButton';
+export { default as NetworkRule } from './NetworkRule';
